#1354: split transmit_prepared_queue into stage/rewrite/verify/write/finalise phases
Move the 230-LOC transmit_prepared_queue() body into a 14-LOC
orchestrator plus five #[inline] phase helpers under a new
transmit/ directory module:

  transmit/mod.rs       — orchestrator + transmit_batch + helpers
                          (TxError, recycle_*, remember_*) and the
                          debug-only log_rst_frames_prepared
  transmit/stage.rs     — pop up to TX_BATCH_SIZE PreparedTxRequests
                          into scratch; orphan-recycle on oversize
  transmit/rewrite.rs   — apply DSCP rewrite to staged UMEM frames
  transmit/verify.rs    — re-validate UMEM slices for staged batch
  transmit/write.rs     — reserve TX ring, write XdpDescs, commit,
                          drop writer, stamp_submits POST-COMMIT
                          (#812 invariant)
  transmit/finalise.rs  — ring-full Retry path or success path
                          (counters + remember_prepared_recycle +
                          retry-tail push-back + maybe_wake_tx)

Pure code motion. Phase ordering preserved 1:1, post-commit submit
stamping (#812) kept inside the write phase by construction so the
orchestrator cannot reorder it, no V_min publish (#940), and the
three orphan-recycle drop sites keep their asymmetric "+= len vs
+= len-1" accounting (#710) verbatim. Zero new per-tick (per-batch)
allocations. All five new phase functions are #[inline].

External callers (queue_service/mod.rs, mirror.rs, tx/drain.rs)
import through the existing tx/mod.rs re-export block; that block
is untouched. Test file moved logically — transmit_tests.rs stays
at its prior path and is loaded by transmit/mod.rs via
#[path = "../transmit_tests.rs"] mod tests so its `use super::*;`
continues to resolve to the orchestrator + private helpers in
transmit/mod.rs.

Plan reviewed by Codex (round-2 to confirm v2) and Gemini round-1
PLAN-READY. Closes #1354.
